refactor(architecture): colocate plugin-owned modules and extract NodeCommands
Move show-completed, tags parsing, and links parsing into their plugin folders; extract NodeCommands to break the plugins/types → OutlineNode type cycle; route completion styling through composeSelfCompleted in picker UIs.
