+++ AGENTS.md
@@ -147,7 +147,7 @@ Feature → seams: **code** A · **links** A+B+I · **route-bible** A(widget)+B 
 
 ## Tag filtering + colors (`src/plugins/tags/`)
 
-`#tags` are **parsed from `node.text`**, never stored. Each renders as a clickable chip (Seam A token); a plain click AND-s that tag into a **URL-driven filter** (`?q=#a #b`) scoped to the zoom `rootId`, re-rendering a **pruned tree** (matches + dimmed ancestor context, everything else hidden). **Filtering is render-time only — it never mutates `collapsed`.** The filter is a Seam-G transform (`buildTagFilter`); the click is Seam-B delegated (`onClick → ctx.nav.filterTag`). Pure logic in `src/data/tags.ts`. `#` autocomplete is the tags plugin's Seam-H menu. v1 is click-driven, tags-only (no free text, no `@`-mentions).
+`#tags` are **parsed from `node.text`**, never stored. Each renders as a clickable chip (Seam A token); a plain click AND-s that tag into a **URL-driven filter** (`?q=#a #b`) scoped to the zoom `rootId`, re-rendering a **pruned tree** (matches + dimmed ancestor context, everything else hidden). **Filtering is render-time only — it never mutates `collapsed`.** The filter is a Seam-G transform (`buildTagFilter`); the click is Seam-B delegated (`onClick → ctx.nav.filterTag`). Pure logic in `src/plugins/tags/tags.ts`. `#` autocomplete is the tags plugin's Seam-H menu. v1 is click-driven, tags-only (no free text, no `@`-mentions).
 
 **Colors** are *chosen* per tag name (not derived) and stored in the `tagColorsCollection` side-collection (Seam E, D1-backed via `/api/kv`) — so they sync and apply to every instance. Painted by **one generated stylesheet** keyed on `data-tag` (`TagColorStyles`, mounted once in `__root.tsx`), so recoloring is an O(1) DOM write with **zero React re-renders**. The picker (`TagColorMenu`) opens on **right-click** (Seam-B `onContextMenu` → `ctx.openOverlay`); the generator skips unsafe tag names (no CSS injection). Why: [Custom tag colors](./docs/DECISIONS.md#custom-tag-colors).
 
@@ -155,7 +155,15 @@ Feature → seams: **code** A · **links** A+B+I · **route-bible** A(widget)+B 
 
 Markdown `[label](url)` **parsed from `node.text`** (Seam A+B+I token), the only construct that **folds**: reveal is **per-link** (Obsidian Live Preview style) — a link shows raw only when the caret is within/adjacent (source offset ∈ `[start, end]`); every other link folds to a clean `<a contenteditable="false">`. At most one reveals at a time.
 
-The landmine: a focused bullet can hold **folded** links, so `el.textContent` is no longer the source. The core is **source-offset-aware** — **`readSource(el)`** (inline-code.ts) reconstructs the markdown (`data-src` for folded `<a>`, `textContent` otherwise) and replaces `el.textContent` in `onInput`/paste **and the slash/tag menus** (else a `/cmd` on a folded-link line drops its url); **`getCaretOffset`/`setCaretOffset`** speak SOURCE offsets, counting a folded link's `data-src-len`. Reveal reflow is a `selectionchange` watcher (`watchCaretReveal`) live only while focused; all of this early-returns on link-free lines (the 99% case). Folded links open on click (Seam-B `window.open`); creation is hand-typed or paste (Seam-I `input.onPaste`, http(s) only, URLs percent-encoded). Search indexes `stripLinks(node.text)`. Why: [Rich links: the source-offset caret](./docs/DECISIONS.md#rich-links-the-source-offset-caret).
+The landmine: a focused bullet can hold **folded** links, so `el.textContent` is no longer the source. The core is **source-offset-aware** — **`readSource(el)`** (inline-code.ts) reconstructs the markdown (`data-src` for folded `<a>`, `textContent` otherwise) and replaces `el.textContent` in `onInput`/paste **and the slash/tag menus** (else a `/cmd` on a folded-link line drops its url); **`getCaretOffset`/`setCaretOffset`** speak SOURCE offsets, counting a folded link's `data-src-len`. Reveal reflow is a `selectionchange` watcher (`watchCaretReveal`) live only while focused; all of this early-returns on link-free lines (the 99% case). Folded links open on click (Seam-B `window.open`); creation is hand-typed or paste (Seam-I `input.onPaste`, http(s) only, URLs percent-encoded). Parsing lives in `plugins/links/links.ts`. Search indexes `stripLinks(node.text)`. Why: [Rich links: the source-offset caret](./docs/DECISIONS.md#rich-links-the-source-offset-caret).
+
+## Todos (`src/plugins/todos/`)
+
+Tasks/completion are the todos plugin's concept — checkboxes, hide-completed, fade cascade, `/todo`/`/bullet`, `[]` autoformat, Mod+Enter/D. The `completed`/`isTask` **fields** stay on `Node` (the one named exception in DECISIONS.md) but all behavior lives here.
+
+- **Show completed** is plugin-owned UI state: `show-completed-provider.tsx` (persisted preference, mounted in `__root.tsx`) and `show-completed-toggle.tsx` (header control — imported explicitly in `Header.tsx` to preserve action order).
+- **Seam G** hide-completed reads `showCompleted` from the provider via `ViewContext`.
+- **`NodeCommands`** (the promoted mutation surface plugins call via `ctx.mutations`) is typed in `src/components/node-commands.ts` — not on `OutlineNode`, so `plugins/types.ts` doesn't import a React component.
 
 ## Daily notes (`src/plugins/daily/`)
