+++ src/components/node-switcher.tsx
@@ -10,8 +10,9 @@ import Fuse, { type FuseResultMatch, type IFuseOptions } from "fuse.js";
 import { Search, BookmarkIcon } from "lucide-react";
 import { useTree } from "../data/useTree";
 import { buildTrail, type Node, type TreeIndex } from "../data/tree";
-import { stripLinks } from "../data/links";
+import { stripLinks } from "../plugins/links/links";
 import {
+  composeSelfCompleted,
   searchAliases,
   searchActions,
   searchAnnotation,
@@ -325,7 +326,7 @@ function ResultRow({
         <span
           className={cn(
             "truncate",
-            node.completed && "text-muted-foreground line-through",
+            composeSelfCompleted(node) && "text-muted-foreground line-through",
           )}
         >
           {highlight(title, textMatchIndices(matches))}
