+++ src/plugins/types.ts
@@ -10,8 +10,8 @@
 import type { ComponentType, ReactNode } from "react";
 import type { Node, TreeIndex } from "../data/tree";
 // Type-only (erased at runtime) -- PluginContext.mutations IS the promoted
-// NodeCommands (D8), so we reference its type without a runtime import cycle.
-import type { NodeCommands } from "../components/OutlineNode";
+// NodeCommands (D8).
+import type { NodeCommands } from "../components/node-commands";
 
 // --- Seam A: inline token + decorator (D6) ---------------------------------
 //
