+++ src/routes/__root.tsx
@@ -6,7 +6,7 @@ import {
   Scripts,
 } from '@tanstack/react-router'
 import { ThemeProvider } from '../components/theme-provider'
-import { ShowCompletedProvider } from '../components/show-completed-provider'
+import { ShowCompletedProvider } from '../plugins/todos/show-completed-provider'
 import { NodeSwitcher } from '../components/node-switcher'
 import { MoveDialog } from '../components/move-dialog'
 import { TagColorStyles } from '../plugins/tags/tag-color-menu'
