+++ src/routes/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { OutlineEditor } from "../components/OutlineEditor";
-import { validateOutlineSearch } from "../data/tags";
+import { validateOutlineSearch } from "../plugins/tags/tags";
 
 export const Route = createFileRoute("/")({
   component: HomePage,
